refactor(dispatcher): decompose monolithic dispatcher.rs into cohesive submodules
Split src/agent/dispatcher.rs (2,593 lines) into focused submodules:

- types.rs: Constants, helper functions, shared types (PREVIEW_MAX_CHARS,
  truncate_for_preview, AgenticLoopResult, PreflightOutcome, ToolBatch,
  check_auth_required, parse_auth_result, etc.)

- delegate.rs: ChatDelegate struct and NativeLoopDelegate implementation,
  including all tool execution helpers (group_tool_calls, execute_one_tool,
  run_tool_batch_inline, run_tool_batch_parallel, etc.)

- core.rs: Agent extension methods (run_agentic_loop, execute_chat_tool)

- tests.rs: All unit tests from the original file

- mod.rs: Thin module root with re-exports preserving existing import paths

Benefits:
- Each submodule has a single, well-defined responsibility
- CodeScene 400-line guideline now satisfied for all non-test files
- No public API breakage - all existing imports continue to work
- Preserved all tracing messages, StatusUpdate payloads, and behavior

Closes #132
